changed the file structure to air b n b standards
